modules/A1-DietRules/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pyyaml==6.0.2
pydantic==2.9.2
